// accounting-system-frontend/lib/schema.d.ts
export interface components {
    schemas: {
        VoucherRequest: {
            /**
             * Format: int32
             * @example 1
             */
            nr: number;
            /**
             * Format: date
             * @example 2023-02-02
             */
            date: string;
            /**
             * Format: int32
             * @example 1920
             */
            account: number;
            /** @example 100 */
            amount: number;
        };
        VoucherResponse: {
            /**
             * Format: int32
             * @example 1
             */
            nr: number;
            /**
             * Format: date
             * @example 2023-02-02
             */
            date: string;
            /**
             * Format: int32
             * @example 1920
             */
            account: number;
            /** @example 100 */
            amount: number;
        };
        Account: {
            /**
             * Format: int32
             * @example 1900
             */
            nr: number;
            /** @example Kontanter */
            name: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}